requirements.txt
jupyter==1.1.1
pandas==2.2.3
seaborn==0.13.2
matplotlib==3.9.2
scikit-learn==1.5.2
geopy==2.4.1
haversine==2.8.1
python-dateutil==2.9.0.post0
sentence-transformers==3.2.1
gender-guesser==0.4.0
scipy==1.14.1
numpy==2.1.3